refactor: move to settings directory
